# dashboard performance enhancements