--- a/mmdvmhost/repeaterinfo.php
+++ b/mmdvmhost/repeaterinfo.php
@@ -66,7 +66,7 @@ if (isset($lastHeard[0])) {
 	                echo "<td style=\"background:#0b0; color:#030;\">Listening</td>";
 	                }
 	        elseif (getActualMode($lastHeard, $mmdvmconfigs) === NULL) {
-	                exec ("pgrep MMDVMHost", $mmdvmhostpid); if (!empty($mmdvmhostpid)) { echo "<td style=\"background:#0b0; color:#030;\">Listening</td>"; } else { echo "<td style=\"background:#606060; color:#b0b0b0;\">OFFLINE</td>"; }
+	                if (isProcessRunning("MMDVMHost")) { echo "<td style=\"background:#0b0; color:#030;\">Listening</td>"; } else { echo "<td style=\"background:#606060; color:#b0b0b0;\">OFFLINE</td>"; }
 	                }
 	        elseif ($listElem[2] && $listElem[6] == null && getActualMode($lastHeard, $mmdvmconfigs) === 'D-Star') {
 	                echo "<td style=\"background:#4aa361;\">RX D-Star</td>";