Some performance enhancements so that pages are faster and use less cpu resources at server. The changes make considerable difference mainly when there are multiple DV modes enabled simultaneously (more pgrep checks).

Benchmark of request times before and after the changes:
repeaterinfo.php: 1.6s -> 436ms
system.php: 1.2s -> 330ms
bm_links.php: 843ms -> 534ms
admin/: 3.49s -> 1.60s
(measured on RPi2, with all DV modes enabled, guess it should make even more difference on RPi0...)