@AndyTaylorTweet 2018-04-24
OK so the only place I have any concern is the life of the static var for the process list, that will "persist for the length of time the script is running" - I'm sure you will have tested it - so I don't have a problem with it - we can always roll back later if its broken anyway :) - so lets give it a go...