@AndyTaylorTweet 2018-04-24
Testing looks good - my concerns about the life of the static array were proven not to be any kind of issue, excellent set of patches, testing on a Pi Zero at the moment;

Before the update, the Zero is totally CPU bound with one user using the dash (testing with a full day of logs here) - load avg is 4.00+, just totally CPU bound - after the update (and a reload of the page in the browser) the system is much happier, load now ~1.20/~1.40 (so still CPU bound but not as badly).