@narspt 2018-04-24
Great, thanks for merging. I have been testing these changes for some days already, think should be all ok, but if I made some mistake we should find it soon :)

I have not a RPi0 myself to test, but I did looked at a friend's one some days ago and CPU usage was really high (with DMR, D-Star and YSF enabled), load average near 5 and CPU idle 0% on 'top'... also his Admin page was really slow to open (but not sure if it was the BM request that was delaying it at that day...)... hope the changes will make things better.